+++ docs/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `with_decision` option for the score function
 - `columns` option for the unzip function
+- support for Python 3.14
 
 ## [3.3.1] - 2025-12-10